Make website sortable items (for ranking) drag & drop-able.
Previously they had move up / move down buttons which have been removed to make the UX cleaner, there are drag handles so that it is still clear what to do.

Note: the dnd-kit sortable library requires that the indexes are non-zero but our API expects zero based indexes so there is a some adding/removing 1 to handle that.